% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPCBrainAge.R
\name{calcPCBrainAge}
\alias{calcPCBrainAge}
\title{calcPCBrainAge}
\usage{
calcPCBrainAge(DNAm, pheno = NULL, CpGImputation = NULL)
}
\arguments{
\item{DNAm}{a matrix of methylation beta values. Needs to be rows = samples and columns = CpGs, with (ideally) rownames and (required) colnames.}

\item{pheno}{Optional: The sample phenotype data (also with samples as rows) that the clock will be appended to.}

\item{CpGImputation}{An optional named vector for the mean value of each CpG that will be input from another dataset if such values are missing here. If no vector is provided, will use data("imputeMissingBrainCpGs") for mean imputation. See data description for more details.}
}
\value{
If you added the optional pheno input (preferred) the function appends a column with the PCBrainAge calculation and returns the dataframe. Otherwise, it will return a vector of calculated clock values in order of input samples.
}
\description{
A function to calculate the predictor of PCBrainAge
}
\examples{
calcPCBrainAge(exampleBetas, examplePheno)
}
