% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeMissingBrainCpGs.R
\docType{data}
\name{imputeMissingBrainCpGs}
\alias{imputeMissingBrainCpGs}
\title{Imputation Data for Missing PCBrainAge CpGs}
\format{
A named vector of length 357,852
}
\source{
<https://doi.org/10.1038/nn.4181>
}
\usage{
imputeMissingBrainCpGs
}
\description{
The mean values of PCBrainAge CpGs obtained from the training data (age > 20) from GSE74193.
}
\keyword{datasets}
