% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanImpute.R
\name{meanImpute}
\alias{meanImpute}
\title{meanImpute}
\usage{
meanImpute(x)
}
\arguments{
\item{x}{A vector which may contain NA values}
}
\value{
A vector where the NA entries have been replaced by the vector's mean
}
\description{
meanImpute
}
\examples{
imputedBetas <- apply(exampleBetas, 2, meanImpute)
}
